`struct Dav1dTileState`: Deduplicate.
